% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_boxplot_prism}
\alias{geom_boxplot_prism}
\title{Prism-style boxplot}
\usage{
geom_boxplot_prism(
  mapping = NULL,
  data = NULL,
  width = 0.6,
  outlier.shape = NA,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings.}

\item{data}{The data to be displayed.}

\item{width}{Box width (default 0.6).}

\item{outlier.shape}{Shape of outliers (\code{NA} = hidden, Prism default).}

\item{...}{Other arguments passed to \code{\link[ggplot2]{geom_boxplot}}.}
}
\description{
A wrapper around \code{\link[ggplot2]{geom_boxplot}} with Prism defaults:
black border, compact width, hidden outliers.  Intended to be paired with
\code{ggchoice("prism")} for authentic GraphPad Prism appearance.
}
\examples{
library(ggplot2)
ggplot(mpg, aes(class, hwy)) +
  geom_boxplot_prism(aes(fill = class)) +
  ggchoice("prism")
}
