% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple Plots}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects can be passed in}

\item{plotlist}{As a list of ggplot objects}

\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\value{
Plot object
}
\description{
Arrange multiple ggplot objects in a grid with a specific number of columns
}
\note{
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
  then plot 1 will go in the upper left, 2 will go in the upper right, and 3
  will go all the way across the bottom.
}
\references{
\url{http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/}
}

