% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_continuous.R
\name{scale_color_matlab_c}
\alias{scale_color_matlab_c}
\alias{scale_fill_matlab_c}
\title{MATLAB parula-style continuous colour scale}
\usage{
scale_color_matlab_c(...)

scale_fill_matlab_c(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2]{scale_colour_gradientn}}.}
}
\description{
Perceptually uniform blue-green-yellow gradient approximating
MATLAB R2014b+ default parula colormap (9 keypoints).
}
