% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_continuous.R
\name{scale_color_origin_c}
\alias{scale_color_origin_c}
\alias{scale_fill_origin_c}
\title{OriginPro-style continuous colour scale}
\usage{
scale_color_origin_c(...)

scale_fill_origin_c(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2]{scale_colour_gradientn}}.}
}
\description{
Blue-to-Yellow gradient via green, mimicking Origin 2017+
"BlueGreenYellow" 2D colormap.
}
