% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_continuous.R
\name{scale_color_prism_c}
\alias{scale_color_prism_c}
\alias{scale_fill_prism_c}
\title{Prism-style continuous colour scale}
\usage{
scale_color_prism_c(...)

scale_fill_prism_c(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2]{scale_colour_gradientn}}.}
}
\description{
Blue-to-red sequential gradient mimicking Prism's default
heatmap colormap. Use for continuous colour aesthetics.
}
