% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_shape_software}
\alias{scale_shape_software}
\alias{scale_shape_prism}
\alias{scale_shape_spss}
\alias{scale_shape_origin}
\alias{scale_shape_stata}
\alias{scale_shape_academic}
\alias{scale_shape_sigmaplot}
\alias{scale_shape_jmp}
\alias{scale_shape_matlab}
\alias{scale_shape_minitab}
\alias{scale_shape_medcalc}
\title{Software-specific point shape scales}
\usage{
scale_shape_prism(...)

scale_shape_spss(...)

scale_shape_origin(...)

scale_shape_stata(...)

scale_shape_academic(...)

scale_shape_sigmaplot(...)

scale_shape_jmp(...)

scale_shape_matlab(...)

scale_shape_minitab(...)

scale_shape_medcalc(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2]{discrete_scale}}.}
}
\description{
Discrete shape scales that apply each graphing software's default
point marker sequence.  Each scale maps shape aesthetics to the
marker shapes used by that software's factory defaults.
}
