% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiPlot.R
\name{multiPlot}
\alias{multiPlot}
\title{multi line plotly}
\usage{
multiPlot(config)
}
\arguments{
\item{config}{a data.frame of length n
the config data.frame must contain a files column which holds the location of the file, a x column which tells the column name to be used for the x axis, and y column for y axis. the data.frame may also contain a trace_names column for the legend.}
}
\value{
pp a plot form plotly
}
\description{
Reading a DF of config with path of the csv file with data, the x column, y column n, legend and color for the line.
}
\author{
Prabhu Manickavelu
}

