% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiPlot.R
\name{multiPlot}
\alias{multiPlot}
\title{multi line plotly}
\usage{
multiPlot(config)
}
\arguments{
\item{config}{a data.frame of length n}
}
\value{
pp a plot form plotly
}
\description{
Reading a DF of config with path of the csv file with data, the x column, y column n, legend and color for the line.
}
\author{
Prabhu Manickavelu
}

