% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/multiplot.r
\name{multiplot}
\alias{multiplot}
\title{A multiplot function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{list of comma separated things to plot}
}
\description{
This function allows the user to plot multiple plots on the same page
}
\examples{
multiplot()
}
\keyword{plot}

